+++ hotbox_designer/designer/menu.py
@@ -12,6 +12,8 @@ class MenuWidget(QtWidgets.QWidget):
     pasteStyleRequested = QtCore.Signal()
     libraryRequested = QtCore.Signal()
     saveToLibraryRequested = QtCore.Signal()
+    testRequested = QtCore.Signal()
+    radialRequested = QtCore.Signal()
     undoRequested = QtCore.Signal()
     redoRequested = QtCore.Signal()
     sizeChanged = QtCore.Signal()
@@ -50,6 +52,14 @@ class MenuWidget(QtWidgets.QWidget):
         self.pastestyle.setToolTip(
             'Paste style on selection, choosing options (Ctrl+Shift+V)')
         self.pastestyle.triggered.connect(self.pasteStyleRequested.emit)
+
+        self.test = QtWidgets.QAction(icon('play.png'), '', self)
+        self.test.setToolTip('Test the hotbox (show it like in production)')
+        self.test.triggered.connect(self.testRequested.emit)
+        self.radial = QtWidgets.QAction(icon('center.png'), '', self)
+        self.radial.setToolTip(
+            'Arrange selected buttons in a circle (marking menu)')
+        self.radial.triggered.connect(self.radialRequested.emit)
 
         self.undo = QtWidgets.QAction(icon('undo.png'), '', self)
         self.undo.setToolTip('Undo')
@@ -165,6 +175,8 @@ class MenuWidget(QtWidgets.QWidget):
             partial(self.arrangeRequested.emit, 'vertical'))
 
         self.toolbar = QtWidgets.QToolBar()
+        self.toolbar.addAction(self.test)
+        self.toolbar.addSeparator()
         self.toolbar.addAction(self.delete)
         self.toolbar.addAction(self.copy)
         self.toolbar.addAction(self.paste)
@@ -201,6 +213,7 @@ class MenuWidget(QtWidgets.QWidget):
         self.toolbar.addSeparator()
         for action in self.align_actions:
             self.toolbar.addAction(action)
+        self.toolbar.addAction(self.radial)
         self.toolbar.addSeparator()
         self.toolbar.addAction(self.arrange_h)
         self.toolbar.addAction(self.arrange_v)